refactor: directory 분리
